requirements.txt
asgiref==3.2.3
Django==3.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
Pillow==6.2.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
